XW-1725 | add tracking to all links in footer
