+++ server/app/views/_partials/global-footer/global-footer-bottom-bar.hbs
@@ -8,12 +8,13 @@
 				model.licensing_and_vertical.description.params.license.title.value
 				tagName='a'
 				href=model.licensing_and_vertical.description.params.license.href
+				trackingLabel='global-footer-licensing-link'
 			)
 			ns='design-system'
 		}}}
 	</div>
 	<div class="wds-global-footer__bottom-bar-row wds-has-border-top global-footer-full-site-link-wrapper">
-		<a id="global-footer-full-site-link" href="?useskin=oasis" rel="nofollow">
+		<a id="global-footer-full-site-link" href="?useskin=oasis" rel="nofollow" data-tracking-label="global-footer-full-site-link">
 			{{i18n 'global-footer-full-site-link' ns='design-system'}}
 		</a>
 	</div>
