+++ server/app/views/_partials/global-footer/global-footer-link-branded.hbs
@@ -1,4 +1,4 @@
-<a href="{{model.href}}" class="wds-global-footer__link wds-is-{{model.brand}}">
+<a href="{{model.href}}" class="wds-global-footer__link wds-is-{{model.brand}}" data-tracking-label="{{model.title.key}}">
 	<div>{{i18n model.title.key ns='design-system'}}</div>
 	{{> svg class='wds-global-footer__image wds-icon' image='wds-icons-arrow'}}
 </a>
