+++ server/app/views/_partials/global-footer/global-footer-link-image.hbs
@@ -1,3 +1,3 @@
-<a href="{{model.href}}" class="wds-global-footer__link">
+<a href="{{model.href}}" class="wds-global-footer__link" data-tracking-label="{{model.title.key}}">
 	{{> svg class='wds-global-footer__image wds-icon' image=model.image}}
 </a>
