+++ server/app/views/_partials/global-footer/global-footer-link-text.hbs
@@ -1,1 +1,1 @@
-<a href="{{model.href}}" class="wds-global-footer__link">{{i18n model.title.key ns='design-system'}}</a>
+<a href="{{model.href}}" class="wds-global-footer__link" data-tracking-label="{{model.title.key}}">{{i18n model.title.key ns='design-system'}}</a>
